Second PR functionality improvements and css changes
