+++ package.json
@@ -31,6 +31,7 @@
   "dependencies": {
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.58",
     "axios": "^0.21.1",
     "canned": "^0.3.12",
     "react": "^17.0.2",
