--- src/Components/App.scss
@@ -1,5 +1,0 @@
-.copyright {
-  text-align: center;
-  padding: 0.5rem;
-  background: #0000000d;
-}
