+++ src/Components/App/App.scss
@@ -1,0 +1,9 @@
+@import url("https://fonts.googleapis.com/css?family=Dosis:400,600,700");
+.copyright {
+  text-align: center;
+  padding: 0.5rem;
+  background: #0000000d;
+}
+* {
+  font-family: "Dosis";
+}
