+++ src/Components/Banner/Banner.js
@@ -1,12 +1,11 @@
 import React, { useState, useEffect } from "react";
-import axios from "axios";
 import { connect } from "react-redux";
-import { fetchBanners } from "../redux/Banners/bannersactions";
-import Eachbanner from "./Eachbanner.js";
+import { fetchBanners } from "../../redux/Banners/bannersactions";
+import Eachbanner from "../Eachbanner/Eachbanner";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.scss";
 import "slick-carousel/slick/slick-theme.scss";
-import "./Banner.scss";
+import "../Banner/Banner.scss";
 function Banner({ bannersData, fetchBanners }) {
   useEffect(() => {
     fetchBanners();
