+++ src/Components/Banner/Banner.scss
