+++ src/Components/Cart/Cart.scss
@@ -5,6 +5,7 @@
   position: absolute;
   width: 400px;
   min-height: 80%;
+  max-height: 85%;
   height: auto;
   background: white;
   top: 50%;
@@ -23,6 +24,8 @@
 .allcartproducts {
   padding: 0.5rem;
   flex: 1;
+  background-color: #8080801a;
+  overflow: auto;
 }
 .emptycarttext {
   text-align: center;
@@ -32,4 +35,19 @@
   padding: 1rem 0;
   background: #bf3737;
   text-align: center;
+}
+.lowestpricebanner {
+  display: flex;
+  align-items: center;
+  justify-content: space-evenly;
+  padding: 0.5rem;
+  background: white;
+  margin-top: 0.5rem;
+}
+.checkout {
+  background: #b81b23;
+  color: white;
+  padding: 1rem;
+  display: flex;
+  justify-content: space-between;
 }
