+++ src/Components/Eachbanner/Eachbanner.scss
