+++ src/Components/Eachcartitem/Eachcartitem.scss
@@ -1,6 +1,9 @@
 .eachcartitem {
   height: 80px;
   display: flex;
+  padding: 0.3rem;
+  background: white;
+  margin-bottom: 0.5rem;
 }
 .eachcartitem img {
   width: 80px;
