+++ src/Components/Eachcategory/Eachcategory.scss
@@ -1,6 +1,3 @@
-* {
-  font-family: sans-serif;
-}
 .eachcategorybutton {
   padding: 0.5rem;
   color: white;
