+++ src/Components/Eachcategory/Eachcategorystyles.js
