+++ src/Components/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Banner from "./Banner";
+import Banner from "./Banner/Banner";
 import Categories from "./Categories";
 function Home() {
   return (
