+++ src/Components/Navbar/Navbar.scss
@@ -3,6 +3,10 @@
   align-items: center;
   height: 100px;
   box-shadow: 0px 8px 11px 0px #00000014;
+  position: sticky;
+  top: 0;
+  background: white;
+  z-index: 11;
 }
 .Header div:first-child {
   display: flex;
@@ -12,6 +16,7 @@
   display: flex;
   align-items: center;
   background: #9e9e9e2e;
+  cursor: pointer;
 }
 .cart {
   height: 40px;
@@ -26,7 +31,6 @@
 }
 .navlinks a {
   width: 80px;
-  font-family: sans-serif;
 }
 @media screen and (min-width: 600px) {
   .Header {
