+++ src/Components/Productlandingpage/Productlandingpage.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import Sidebar from "./Sidebar";
-import Products from "./Products";
+import Sidebar from "../Sidebar/Sidebar";
+import Products from "../Products/Products";
 import { connect } from "react-redux";
 import "./Productlandingpage.scss";
-import { fetchProducts } from "../redux/Products/productsaction";
-import { fetchCategories } from "../redux/Categories/categoriesactions";
+import { fetchProducts } from "../../redux/Products/productsaction";
+import { fetchCategories } from "../../redux/Categories/categoriesactions";
 function Productlandingpage({
   productsData,
   categoriesData,
@@ -24,9 +24,9 @@ function Productlandingpage({
   }, []);
 
   const eventhandler = (data) => {
-    const filtereddata = productsData.products.filter(
-      (item) => item.category == data
-    );
+    const filtereddata = data
+      ? productsData.products.filter((item) => item.category == data)
+      : productsData.products;
     console.log(filtereddata, "filtereddata");
     setcurrentproducts(filtereddata);
     setrerender(true);
