+++ src/index.js
@@ -1,4 +1,4 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./Components/App.js";
+import App from "./Components/App/App.js";
 ReactDOM.render(<App />, document.getElementById("app"));
